Check ob Schreibrechte im Migrationsordner vorhanden

+++ layouts/db/mysql/schema/2020-11-06_10-00-00_migrate_to_3.0.php
@@ -7,31 +7,28 @@
 	$this->database->execSQL($sql, 0, 0);
 
 	if ($this->database->result->num_rows > 0) {
-		echo '<br>Anzahl der Einträge in migrations Tabelle: ' . $this->database->result->num_rows;
-		while ($rs = $this->database->result->fetch_array()) {
-			echo '<br>Komponente: ' . $rs['component'] . ' Typ: ' . $rs['type'] . ' Datei: ' . $rs['filename'];
-		}
-		echo '<br>Download alle Migrations-Dateien von vor Version 3.0 um eventuell fehlende auszuführen.';
 		$success = false;
-		$zip_file = LAYOUTPATH.'db/mysql/schema/mysql_migrations_upto_3.0.zip';
-		$zip_url = 'https://gdi-service.de/public/kvwmap_resources/mysql_migrations_upto_3.0.zip';
-		stream_context_set_default(
-			array(
-				'ssl' => array(
-					'verify_peer'      => FALSE,
-					'verify_peer_name' => FALSE
+		if(!is_writable(LAYOUTPATH.'db/mysql/schema')){
+			$GUI->add_message('Keine Schreibrechte in '.LAYOUTPATH.'db/mysql/schema/. Sorgen Sie bitte in diesem Verzeichnis für Gruppenschreibrechte.');
+		}
+		else{
+			$zip_file = LAYOUTPATH.'db/mysql/schema/mysql_migrations_upto_3.0.zip';
+			$zip_url = 'https://gdi-service.de/public/kvwmap_resources/mysql_migrations_upto_3.0.zip';
+			stream_context_set_default(
+				array(
+					'ssl' => array(
+						'verify_peer'      => FALSE,
+						'verify_peer_name' => FALSE
+					)
 				)
-			)
-		);
-		file_put_contents($zip_file, file_get_contents($zip_url));
-		if (file_exists($zip_file)) {
-			unzip($zip_file, false, false, true);
-			unlink($zip_file);
-			$success = true;
+			);
+			file_put_contents($zip_file, file_get_contents($zip_url));
+			if (file_exists($zip_file)) {
+				unzip($zip_file, false, false, true);
+				unlink($zip_file);
+				$success = true;
+			}
 		}
 		$result[0] = !$success;
-	}
-	else {
-		echo '<br>Keine registrierten Migrationsdateien gefunden.';
 	}
 ?>
